# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/teachers": {
      "get": {
        "operationId": "adminListTeachers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeacherWithCoursesDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "List all teachers",
        "tags": [
          "admin: teachers"
        ]
      },
      "post": {
        "operationId": "adminCreateTeacher",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherWithCoursesDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Create a teacher",
        "tags": [
          "admin: teachers"
        ]
      }
    },
    "/api/v1/admin/teachers/{id}": {
      "get": {
        "operationId": "adminGetTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherWithCoursesDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Get a teacher",
        "tags": [
          "admin: teachers"
        ]
      },
      "patch": {
        "operationId": "adminUpdateTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherWithCoursesDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed; `errors` maps field paths to messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Update a teacher",
        "tags": [
          "admin: teachers"
        ]
      },
      "delete": {
        "operationId": "adminDeleteTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Teacher is still assigned to a course or batch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "workos": []
          }
        ],
        "summary": "Delete a teacher",
        "tags": [
          "admin: teachers"
        ]
      }
    },
...
  },
  "info": {
    "title": "PrincetonBD API",
    "description": "Backend API for the PrincetonBD web application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "workos": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "WorkOS AuthKit access token."
      }
    },
    "schemas": {
...
      "BranchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "dhaka-gulshan"
          },
          "name": {
            "type": "string",
            "example": "Dhaka — Gulshan"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "example": "House 00, Road 00, Gulshan-2, Dhaka 1212"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+880 1700-000000"
          },
          "isActive": {
            "type": "boolean",
            "description": "Inactive branches are hidden from public lists but keep their history."
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "address",
          "phone",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Course not found"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "name": [
                "name should not be empty"
              ],
              "modules.0.title": [
                "title must be a string"
              ]
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "errors"
        ]
      },
...
      "CourseStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ]
      },
      "CourseRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "sat"
          },
          "name": {
            "type": "string",
            "example": "SAT"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "status"
        ]
      },
      "TeacherWithCoursesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "farzana-haque"
          },
          "name": {
            "type": "string",
            "example": "Farzana Haque"
          },
          "designation": {
            "type": "string",
            "description": "The title shown under the name.",
            "example": "Lead Instructor, Quantitative"
          },
          "bio": {
            "type": "string",
            "description": "One or two lines: the credential that matters."
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "description": "Public URL of an uploaded photo; null renders a monogram."
          },
          "branch": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BranchDto"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "courses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseRefDto"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "designation",
          "bio",
          "imageUrl",
          "branch",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt",
          "courses"
        ]
      },
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL segment. Derived from `name` when omitted.",
            "maxLength": 80,
            "example": "farzana-haque",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "Farzana Haque"
          },
          "designation": {
            "type": "string",
            "maxLength": 120,
            "example": "Lead Instructor, Quantitative"
          },
          "bio": {
            "type": "string",
            "maxLength": 1000
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 1024,
            "format": "uri"
          },
          "branchId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "name",
          "designation"
        ]
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "URL segment. Derived from `name` when omitted.",
            "maxLength": 80,
            "example": "farzana-haque",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "Farzana Haque"
          },
          "designation": {
            "type": "string",
            "maxLength": 120,
            "example": "Lead Instructor, Quantitative"
          },
          "bio": {
            "type": "string",
            "maxLength": 1000
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 1024,
            "format": "uri"
          },
          "branchId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        }
      },
...
    }
  }
}